Move LogStore implementation into own module
