+++ p2panda-rs/src/storage_provider/traits/log_store.rs
@@ -56,89 +56,12 @@ pub trait LogStore<StorageLog: AsStorageLog> {
 pub mod tests {
     use std::convert::TryFrom;
 
-    use async_trait::async_trait;
     use rstest::rstest;
 
-    use crate::document::DocumentId;
     use crate::entry::LogId;
-    use crate::identity::{Author, KeyPair};
-    use crate::schema::SchemaId;
-    use crate::storage_provider::errors::LogStorageError;
-    use crate::storage_provider::traits::test_utils::{
-        test_db, SimplestStorageProvider, StorageLog, TestStore,
-    };
-    use crate::storage_provider::traits::{AsStorageLog, LogStore};
-    use crate::test_utils::fixtures::{document_id, key_pair, schema};
-
-    /// Implement the `LogStore` trait on SimplestStorageProvider
-    #[async_trait]
-    impl LogStore<StorageLog> for SimplestStorageProvider {
-        async fn insert_log(&self, log: StorageLog) -> Result<bool, LogStorageError> {
-            self.db_insert_log(log);
-            Ok(true)
-        }
-
-        /// Get a log from storage
-        async fn get(
-            &self,
-            author: &Author,
-            document_id: &DocumentId,
-        ) -> Result<Option<LogId>, LogStorageError> {
-            let logs = self.logs.lock().unwrap();
-
-            let log = logs
-                .iter()
-                .find(|log| log.document_id() == *document_id && log.author() == *author);
-
-            let log_id = log.map(|log| log.id());
-            Ok(log_id)
-        }
-
-        async fn next_log_id(&self, author: &Author) -> Result<LogId, LogStorageError> {
-            let logs = self.logs.lock().unwrap();
-
-            let author_logs = logs.iter().filter(|log| log.author() == *author);
-            let next_log_id = author_logs.count() + 1;
-            Ok(LogId::new(next_log_id as u64))
-        }
-    }
-
-    #[rstest]
-    #[async_std::test]
-    async fn insert_get_log(key_pair: KeyPair, schema: SchemaId, document_id: DocumentId) {
-        // Instantiate a new store.
-        let store = SimplestStorageProvider::default();
-
-        let author = Author::try_from(key_pair.public_key().to_owned()).unwrap();
-        let log = StorageLog::new(&author, &schema, &document_id, &LogId::default());
-
-        // Insert a log into the store.
-        assert!(store.insert_log(log).await.is_ok());
-
-        // Get a log_id from the store by author and document_id.
-        let log_id = store.get(&author, &document_id).await;
-
-        assert!(log_id.is_ok());
-        assert_eq!(log_id.unwrap().unwrap(), LogId::default())
-    }
-
-    #[rstest]
-    #[async_std::test]
-    async fn get_next_log_id(key_pair: KeyPair, schema: SchemaId, document_id: DocumentId) {
-        // Instantiate a new store.
-        let store = SimplestStorageProvider::default();
-
-        let author = Author::try_from(key_pair.public_key().to_owned()).unwrap();
-        let log_id = store.next_log_id(&author).await.unwrap();
-        assert_eq!(log_id, LogId::default());
-
-        let log = StorageLog::new(&author, &schema, &document_id, &LogId::default());
-
-        assert!(store.insert_log(log).await.is_ok());
-
-        let log_id = store.next_log_id(&author).await.unwrap();
-        assert_eq!(log_id, LogId::new(2));
-    }
+    use crate::identity::Author;
+    use crate::storage_provider::traits::test_utils::{test_db, TestStore};
+    use crate::storage_provider::traits::LogStore;
 
     #[rstest]
     #[async_std::test]
