+++ p2panda-rs/src/test_utils/db/stores/mod.rs
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod entry;
+mod log;
 mod operation;
